refactor: move fs helpers in utils package
