+++ pkg/openapiclient/overlay.go
@@ -10,6 +10,7 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/openapi"
+	"sigs.k8s.io/kubectl-validate/pkg/utils"
 )
 
 //go:embed patches
@@ -26,11 +27,11 @@ func PatchLoaderFromDirectory(filesystem fs.FS, dir string) PatchLoaderFn {
 		return nil
 	}
 	return func(s string) []byte {
-		if res, err := readFile(filesystem, filepath.Join(dir, s+".json")); err == nil {
+		if res, err := utils.ReadFile(filesystem, filepath.Join(dir, s+".json")); err == nil {
 			return res
-		} else if res, err := readFile(filesystem, filepath.Join(dir, s+".yaml")); err == nil {
+		} else if res, err := utils.ReadFile(filesystem, filepath.Join(dir, s+".yaml")); err == nil {
 			return res
-		} else if res, err := readFile(filesystem, filepath.Join(dir, s+".yml")); err == nil {
+		} else if res, err := utils.ReadFile(filesystem, filepath.Join(dir, s+".yml")); err == nil {
 			return res
 		}
 		return nil
